+++ app/src/main/java/com/example/foodie_finder/data/local/AppLocalDb.kt
@@ -1,12 +1,12 @@
-package com.example.foodie_finder.model.dao
+package com.example.foodie_finder.data.local
 
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import com.example.foodie_finder.base.MyApplication
-import com.example.foodie_finder.model.Converters
-import com.example.foodie_finder.model.Student
+import com.example.foodie_finder.data.local.dao.StudentDao
+import com.example.foodie_finder.data.model.Converters
 
 @Database(entities = [Student::class], version = 1)
 @TypeConverters(Converters::class)
