+++ app/src/main/java/com/example/foodie_finder/data/local/Student.kt
@@ -1,8 +1,7 @@
-package com.example.foodie_finder.model
+package com.example.foodie_finder.data.local
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.google.firebase.Timestamp
 import java.sql.Time
 import java.util.Date
