+++ app/src/main/java/com/example/foodie_finder/data/local/dao/StudentDao.kt
@@ -1,4 +1,4 @@
-package com.example.foodie_finder.model.dao
+package com.example.foodie_finder.data.local.dao
 
 import androidx.room.Dao
 import androidx.room.Delete
@@ -6,7 +6,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.example.foodie_finder.model.Student
+import com.example.foodie_finder.data.local.Student
 
 @Dao
 interface StudentDao {
