+++ app/src/main/java/com/example/foodie_finder/data/remote/CloudinaryModel.kt
@@ -1,4 +1,4 @@
-package com.example.foodie_finder.model
+package com.example.foodie_finder.data.remote
 
 import android.content.Context
 import android.graphics.Bitmap
@@ -10,6 +10,7 @@ import com.cloudinary.android.policy.GlobalUploadPolicy
 import com.cloudinary.android.policy.UploadPolicy
 import com.example.foodie_finder.BuildConfig
 import com.example.foodie_finder.base.MyApplication
+import com.example.foodie_finder.utils.extentions.toFile
 import java.io.File
 import java.io.FileOutputStream
 
@@ -33,12 +34,12 @@ class CloudinaryModel {
 
     fun uploadBitmap(
         bitmap: Bitmap,
-        name: String?,
+        name: String,
         onSuccess: (String) -> Unit,
         onError: (String) -> Unit
     ) {
         val context = MyApplication.Globals.context ?: return
-        val file = bitmapToFile(bitmap, name, context)
+        val file = bitmap.toFile(context,name)
 
         MediaManager.get().upload(file.path)
             .option(
@@ -81,20 +82,20 @@ class CloudinaryModel {
             .dispatch()
     }
 
-    private fun bitmapToFile(
-        bitmap: Bitmap,
-        name: String?,
-        context: Context
-    ): File {
-        val file =
-            File(context.cacheDir, "${name ?: "temp_image_${System.currentTimeMillis()}"}.jpg")
-        try {
-            FileOutputStream(file).use { outputStream ->
-                bitmap.compress(Bitmap.CompressFormat.JPEG, 100, outputStream)
-            }
-        } catch (e: Exception) {
-            Log.e("CloudinaryModel", "Error saving bitmap to file", e)
-        }
-        return file
-    }
+//    private fun bitmapToFile(
+//        bitmap: Bitmap,
+//        name: String?,
+//        context: Context
+//    ): File {
+//        val file =
+//            File(context.cacheDir, "${name ?: "temp_image_${System.currentTimeMillis()}"}.jpg")
+//        try {
+//            FileOutputStream(file).use { outputStream ->
+//                bitmap.compress(Bitmap.CompressFormat.JPEG, 100, outputStream)
+//            }
+//        } catch (e: Exception) {
+//            Log.e("CloudinaryModel", "Error saving bitmap to file", e)
+//        }
+//        return file
+//    }
 }
