+++ app/src/main/java/com/example/foodie_finder/ui/view/EditPostFragment.kt
@@ -1,4 +1,4 @@
-package com.example.foodie_finder
+package com.example.foodie_finder.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,12 +6,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.Navigation
+import com.example.foodie_finder.R
 import com.example.foodie_finder.databinding.FragmentEditPostBinding
-import com.example.foodie_finder.dialogs.alert.showSuccessOperationDialog
-import com.example.foodie_finder.model.Model
-import com.example.foodie_finder.model.Student
-import com.example.foodie_finder.dialogs.pickers.showDatePickerDialog
-import com.example.foodie_finder.dialogs.pickers.showTimePickerDialog
+import com.example.foodie_finder.utils.dialogs.alert.showSuccessOperationDialog
+import com.example.foodie_finder.data.model.Model
+import com.example.foodie_finder.data.local.Student
+import com.example.foodie_finder.utils.dialogs.pickers.showDatePickerDialog
+import com.example.foodie_finder.utils.dialogs.pickers.showTimePickerDialog
 import com.example.foodie_finder.utils.DateTimeUtils
 import java.sql.Time
