+++ app/src/main/java/com/example/foodie_finder/ui/view/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.example.foodie_finder
+package com.example.foodie_finder.ui.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,28 +7,34 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
-import com.example.foodie_finder.databinding.FragmentRegisterBinding
-import com.example.foodie_finder.model.Model
+import com.example.foodie_finder.LoginFragmentDirections
+import com.example.foodie_finder.databinding.FragmentLoginBinding
+import com.example.foodie_finder.data.model.Model
 
-class RegisterFragment : Fragment() {
+class LoginFragment : Fragment() {
 
-    private var binding: FragmentRegisterBinding? = null
-
+    private var binding: FragmentLoginBinding? = null
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentRegisterBinding.inflate(inflater, container, false)
-        binding?.loginButton?.setOnClickListener {
-            binding?.root?.let { Navigation.findNavController(it).popBackStack() }
-        }
-
-        binding?.signUpButton?.setOnClickListener {
-            signUpAction()
-        }
-
+        binding = FragmentLoginBinding.inflate(inflater, container, false)
         return binding?.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding?.loginButton?.setOnClickListener {
+            loginAction()
+        }
+
+        binding?.registerButton?.setOnClickListener {
+            val action =
+                LoginFragmentDirections.actionLoginFragmentToRegisterFragment()
+            binding?.root?.let { Navigation.findNavController(it).navigate(action) }
+        }
     }
 
     override fun onDestroyView() {
@@ -36,16 +42,13 @@ class RegisterFragment : Fragment() {
         binding = null
     }
 
-    private fun signUpAction() {
+    private fun loginAction() {
         val email = binding?.emailEditText?.text.toString().trim()
         val password = binding?.passwordEditText?.text.toString().trim()
-        val passwordConfirm = binding?.confirmPasswordEditText?.text.toString().trim()
-        val firstName = binding?.firstNameEditText?.text.toString().trim()
-        val lastName = binding?.lastNameEditText?.text.toString().trim()
+
 
         binding?.emailInputLayout?.error = null
         binding?.passwordInputLayout?.error = null
-        binding?.confirmPasswordInputLayout?.error = null
 
         var isValid = true
 
@@ -59,21 +62,14 @@ class RegisterFragment : Fragment() {
             isValid = false
         }
 
-        if (passwordConfirm != password) {
-            binding?.confirmPasswordInputLayout?.error = "Passwords do not match"
-            binding?.passwordInputLayout?.error = "Passwords do not match"
-            isValid = false
-        }
-
         if (!isValid) return
 
-        Model.shared.signUp(firstName, lastName, email, password) { success, message, errorFields ->
-            if (success) {
-                Toast.makeText(requireContext(), "Register successful!", Toast.LENGTH_LONG)
-                    .show()
 
+        Model.shared.signIn(email, password) { success, message, errorFields ->
+            if (success) {
+                Toast.makeText(requireContext(), "Login successful!", Toast.LENGTH_LONG).show()
                 val action =
-                    RegisterFragmentDirections.actionRegisterFragmentToStudentsListFragment()
+                    LoginFragmentDirections.actionLoginFragmentToStudentsListFragment()
                 binding?.root?.let { Navigation.findNavController(it).navigate(action) }
             } else {
                 errorFields?.forEach { field ->
@@ -86,7 +82,7 @@ class RegisterFragment : Fragment() {
                 if (errorFields.isNullOrEmpty()) {
                     Toast.makeText(
                         requireContext(),
-                        message ?: "Register failed",
+                        message ?: "Login failed",
                         Toast.LENGTH_LONG
                     ).show()
                 }
