+++ app/src/main/java/com/example/foodie_finder/ui/view/ProfileFragment.kt
@@ -1,10 +1,11 @@
-package com.example.foodie_finder
+package com.example.foodie_finder.ui.view
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.foodie_finder.R
 
 
 class ProfileFragment : Fragment() {
