+++ app/src/main/java/com/example/foodie_finder/ui/viewModel/PostsListViewModel.kt
@@ -1,0 +1,8 @@
+package com.example.foodie_finder.ui.viewModel
+
+import androidx.lifecycle.ViewModel
+import com.example.foodie_finder.data.local.Student
+
+class PostsListViewModel : ViewModel() {
+
+}
