+++ app/src/main/java/com/example/foodie_finder/utils/extentions/BitmapExtentions.kt
@@ -1,0 +1,14 @@
+package com.example.foodie_finder.utils.extentions
+
+import android.content.Context
+import android.graphics.Bitmap
+import java.io.File
+import java.io.FileOutputStream
+
+fun Bitmap.toFile(context: Context, name: String): File {
+    val file = File(context.cacheDir, "image_$name.jpg")
+    FileOutputStream(file).use { stream ->
+        compress(Bitmap.CompressFormat.JPEG, 100, stream)
+    }
+    return file
+}
