+++ app/src/main/res/layout/fragment_edit_post.xml
@@ -4,7 +4,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EditPostFragment">
+    tools:context=".ui.view.EditPostFragment">
 
     <include
         android:id="@+id/student_input_form"
