+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="24dp"
-    tools:context=".LoginFragment">
+    tools:context=".ui.view.LoginFragment">
 
     <TextView
         android:id="@+id/appTitle"
