+++ app/src/main/res/layout/fragment_new_post.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NewPostFragment">
+    tools:context=".ui.view.NewPostFragment">
 
     <include
         android:id="@+id/student_input_form"
