+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.example.foodie_finder.LoginFragment"
+        android:name="com.example.foodie_finder.ui.view.LoginFragment"
         android:label="Login"
         tools:layout="@layout/fragment_login">
         <action
@@ -20,7 +20,7 @@
 
     <fragment
         android:id="@+id/registerFragment"
-        android:name="com.example.foodie_finder.RegisterFragment"
+        android:name="com.example.foodie_finder.ui.view.RegisterFragment"
         android:label="Register"
         tools:layout="@layout/fragment_register">
         <action
@@ -33,7 +33,7 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.example.foodie_finder.HomeFragment"
+        android:name="com.example.foodie_finder.ui.view.HomeFragment"
         tools:layout="@layout/fragment_posts_list">
         <action
             android:id="@+id/action_studentsListFragment_to_studentDetailsFragment"
@@ -54,7 +54,7 @@
 
     <fragment
         android:id="@+id/postDetailsFragment"
-        android:name="com.example.foodie_finder.PostDetailsFragment"
+        android:name="com.example.foodie_finder.ui.view.PostDetailsFragment"
         android:label="Post Details"
         tools:layout="@layout/fragment_post_details">
         <argument
@@ -67,7 +67,7 @@
 
     <fragment
         android:id="@+id/editPostFragment"
-        android:name="com.example.foodie_finder.EditPostFragment"
+        android:name="com.example.foodie_finder.ui.view.EditPostFragment"
         android:label="Edit Student"
         tools:layout="@layout/fragment_edit_post">
         <argument
@@ -77,25 +77,25 @@
 
     <fragment
         android:id="@+id/newPostFragment"
-        android:name="com.example.foodie_finder.NewPostFragment"
+        android:name="com.example.foodie_finder.ui.view.NewPostFragment"
         android:label="Add Student"
         tools:layout="@layout/fragment_new_post" />
 
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.example.foodie_finder.SearchFragment"
+        android:name="com.example.foodie_finder.ui.view.SearchFragment"
         android:label="Search"
         tools:layout="@layout/fragment_search"/>
 
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.example.foodie_finder.ProfileFragment"
+        android:name="com.example.foodie_finder.ui.view.ProfileFragment"
         android:label="Search"
         tools:layout="@layout/fragment_profile"/>
 
     <fragment
         android:id="@+id/savedPostsFragment"
-        android:name="com.example.foodie_finder.SavedPostsFragment"
+        android:name="com.example.foodie_finder.ui.view.SavedPostsFragment"
         android:label="Search"
         tools:layout="@layout/fragment_saved_posts"/>
